requirements.txt
fastapi
langchain-core
langchain-openai
langgraph
pydantic
python-dotenv
tavily-python
uvicorn
